sanity bar modification

+++ IMM460Experiment2/Assets/Scripts/SanitySystem/sanityBar.cs
@@ -9,14 +9,14 @@ public class sanityBar : MonoBehaviour
     public Gradient gradient;
     public Image fill;
 
-    public void SetMaxHealth(int health) {
+    public void SetMaxHealth(float health) {
         slider.maxValue = health;
         slider.value = health;
 
         fill.color = gradient.Evaluate(1f);
     }
 
-    public void SetHealth(int health) {
+    public void SetHealth(float health) {
         slider.value = health;
         fill.color = gradient.Evaluate(slider.normalizedValue);
     }
